feat!: restructure config to be more configurable
