+++ lua/nvim-semantic-tokens.lua
@@ -1,56 +1,9 @@
 local M = {}
+local highlighters = {}
 
 local utils = require "nvim-semantic-tokens.utils"
 local semantic_tokens = require "vim.lsp.semantic_tokens"
 local ns = vim.api.nvim_create_namespace "nvim-semantic-tokens"
-
-M.token_map = {}
-M.modifiers_map = {}
-
---- transforms a highlight group string or a table of highlight group strings
-local function define_prefixed_hl(ft, hl)
-  local prefixed
-  if type(hl) == "table" then
-    prefixed = {}
-    for k, sub_hl in pairs(hl) do
-      prefixed[k] = define_prefixed_hl(ft, sub_hl)
-    end
-  else
-    prefixed = ft .. hl
-    vim.cmd("highlight default link " .. prefixed .. " " .. hl)
-  end
-  return prefixed
-end
-
-local function make_cache(base_map)
-  -- two-stage look-up: ft-> token -> highlight
-  -- two-stage look-up: ft-> modifier -> highlight (or per-token table for three stage look-up)
-  return setmetatable({}, {
-    __index = function(ft_cache, ft)
-      local rtn = setmetatable({}, {
-        __index = function(hl_cache, query)
-          local hl = base_map[query]
-          if hl then
-            local prefixed = define_prefixed_hl(ft, hl)
-            rawset(hl_cache, query, prefixed)
-            return prefixed
-          end
-          return hl
-        end,
-      })
-      rawset(ft_cache, ft, rtn)
-      return rtn
-    end,
-  })
-end
-
-local token_cache
-local modifiers_cache
-
-local function reset_cache()
-  token_cache = make_cache(M.token_map)
-  modifiers_cache = make_cache(M.modifiers_map)
-end
 
 local function highlight(buf, token, hl)
   vim.api.nvim_buf_set_extmark(buf, ns, token.line, token.start_char, {
@@ -63,24 +16,14 @@ local function highlight(buf, token, hl)
 end
 
 local function highlight_token(ctx, token)
-  local buf = ctx.bufnr
-  local ft = vim.api.nvim_buf_get_option(buf, "filetype")
-  local ft_token_cache = token_cache[ft]
-  local ft_modifiers_cache = modifiers_cache[ft]
-
-  local hl = ft_token_cache[token.type]
-  if hl then
-    highlight(buf, token, hl)
+  local function highlight_fn(hl)
+    highlight(ctx.bufnr, token, hl)
   end
-  for _, m in pairs(token.modifiers) do
-    hl = ft_modifiers_cache[m]
-    -- modifiers can have a per-type mapping
-    -- e.g. readonly = { variable = "ReadOnlyVariable" }
-    if type(hl) == "table" then
-      hl = hl[token.type]
-    end
-    if hl then
-      highlight(buf, token, hl)
+  for _, highlighter in ipairs(highlighters) do
+    if type(highlighter) == 'table'then
+      highlighter.highlight_token(ctx, token, highlight_fn)
+    else
+      highlighter(ctx, token, highlight_fn)
     end
   end
 end
@@ -98,7 +41,12 @@ function M.setup(config)
   if config.preset and config.preset ~= "default" then
     vim.cmd("luafile " .. utils.get_preset_file(config.preset))
   end
-  reset_cache()
+  highlighters = config.highlighters or { require "nvim-semantic-tokens.table-highlighter" }
+  for _, h in ipairs(highlighters) do
+    if type(h) == 'table' and h.reset then
+      h.reset()
+    end
+  end
 end
 
 return M
